Connection page added.
Now we have a working connection with google account
Encryption and access are resolves.
https added with self signed cert ( need to find a google acceptable one )

added the open salt function.

using absolute path insted of relative to open encrypted files

NEXT STEP :
we need to find a way to FORCE the user to be connected AND connected to the right account. Because right now they CAN just use someone else account or input anything they want in the URL and make the account they want or access someone else account.

A way to get the password for a user and unlock there files insted of 'test' that we are using right now