Yea. Please read the doc.
If we read the doc and use google correctly we can now use it as it was suppose to. The back up is inside the account of the app insted of my own. I am unsure if its betther because if I loose access to the account I won't see the back up ever again. I could share it with my own account.