start of the implementation of backup
new todo added :
    #TODO:NEED TO ENSURE THE USERS FILES ARE ALL ENCRYPTED

#TODO: Look if the file is encrypted or not. If it is ensure its stay that way up to the end of the backup

temp_user_password: str = 'test' #TODO: Not use test and use some unchangeable value of a google account to create a strong password (maybe?)

DB_FILE_PATH: str = './usr_data/' #TODO: Receive it as an argument

So maybe I don't know. Kinda like make bether password and look if a file is encrypted. The easy way for the json file I have is to see if we do a json load it read it. If not HIGH chance its encrypted