how to keep file
Ok, the safest way to ensure no files are ever backed up when they are unencrypted is to never decrypt them.

I have change two function to only read or write the encrypted version.
This would ensure that even when the user is asking for information if there is a back up we do not by mistake take the decrypted version of the file.

Now we need to implement them where they need to be.

Also added a function to create the derivation key insted of doing it in every function.