website passwords works
Now the user once loggin need to give is own password. Once the password is given the file will decrypt correctly.

this was put in place so the website owner should NOT be able to decrypt the file of one of the user.

The google account is there just to be sur we give the file to the correct user.